+++ R/LTMRpilot.R
@@ -4,6 +4,8 @@
 #' @param quiet Set \code{quiet=TRUE} to suppress the warning message about length data.
 #' @param convert_lengths Should all species with conversion equations have their lengths converted to fork length (or total length if no fork)?
 #' @param remove_unconverted_lengths Should all species without a conversion equation be removed? Ignored if \code{convert_lengths=FALSE}.
+#' @param size_cutoff Minimum fish length (in mm) that should be retained in the data. Any fish below this length will be removed.
+#' It is advised to convert lengths before using this option. All \code{NA} lengths will be retained. Set \code{size_cutoff=NULL} to keep all sizes.
 #' @param measured_lengths Only include measured lengths. Does not calculate adjusted length frequencies. Will not represent total fish caught.
 #'
 #' @importFrom magrittr %>%
@@ -16,7 +18,7 @@
 #'
 #' # All data, convert lengths to Fork length (or Total length if not fork)
 #' # but retain all data, even those that could not be converted
-#' Data <- LTMRpilot(convert_lengths=TRUE, remove_unconverted_lengths=TRUE)
+#' Data <- LTMRpilot(convert_lengths=TRUE, remove_unconverted_lengths=FALSE)
 #'
 #'
 #' # Only outputting measured lengths, without frequencies adjusted for counts of unmeasured fish,
@@ -27,6 +29,7 @@
 LTMRpilot <- function(quiet=FALSE,
                       convert_lengths=TRUE,
                       remove_unconverted_lengths=TRUE,
+                      size_cutoff=NULL,
                       measured_lengths=FALSE){
 
   data<-dplyr::bind_rows(LTMRdata::Suisun, LTMRdata::Baystudy, LTMRdata::FMWT)
@@ -51,12 +54,16 @@ LTMRpilot <- function(quiet=FALSE,
     data<-dplyr::mutate(data, Species=stringr::str_remove(.data$Taxa, " \\((.*)"))
     Length_conversions<-LTMRdata::Length_conversions
     if(remove_unconverted_lengths){
-      data<-dplyr::filter(data, .data$Species%in%unique(Length_conversions$Species))
+      data<-dplyr::filter(data, .data$Species%in%unique(Length_conversions$Species) | is.na(.data$Species))
     }
     data<-data%>%
       dplyr::left_join(Length_conversions, by="Species")%>%
       dplyr::mutate(Length=dplyr::if_else(.data$Source=="Suisun" & .data$Species%in%unique(Length_conversions$Species), .data$Intercept+.data$Slope*.data$Length, .data$Length))%>%
       dplyr::select(-.data$Intercept, -.data$Slope, -.data$Species)
+  }
+
+  if(!is.null(size_cutoff)){
+    data<-dplyr::filter(data, is.na(.data$Length) | .data$Length>=size_cutoff)
   }
 
   if(!quiet){